python/agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.0.1"
description = "Agent project"
authors = ["agent-team"]
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include = "server" },
    { include = "core" },
]

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[[tool.poetry.source]]
name = "pypi"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
uvicorn = { version = "^0.19.0", extras = ["standard"] }
fastapi = "^0.95.0"
orjson = "^3.8.0"
python-dotenv = "^1.0.0"
pypinyin = "^0.50.0"
grandalf = "^0.8"

[tool.poetry.group.agent]
optional = true
[tool.poetry.group.agent.dependencies]
langchain = "^0.1.3"
langchain-openai = "^0.1.1"
langgraph = "^0.0.31"
tavily-python = "^0.3.3"
langchainhub = "^0.1.15"

[tool.poetry.group.test]
optional = true
[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.4"

[tool.poetry-version-plugin]
source = "file"
file_path = "../VERSION"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"